+++ tests/infra/jwt_issuer.py
@@ -158,7 +158,7 @@ class JwtIssuer:
         else:
             raise ValueError(f"Unsupported algorithm: {self._alg}")
 
-        cert = infra.crypto.generate_cert(key_priv, cn=cn, ca=True)
+        cert = infra.crypto.generate_cert(key_priv, cn=cn, ca=True, san=cn)
         return (key_priv, key_pub), cert
 
     def __init__(